pyproject.toml
[project]
name = "monitor-procon"
version = "0.1.0"
description = "Monitor de vagas do Procon MA"
authors = [
    {name = "Matheus",email = "mathii10costa@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
